requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
audioop-lts==0.2.1
babel==2.17.0
bibtexparser==2.0.0b8
blinker==1.9.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
clldutils==3.24.1
colorama==0.4.6
colorlog==6.9.0
csvw==3.5.1
deepgram-sdk==2.12.0
dlinfo==2.0.0
fastapi==0.115.8
ffmpy==0.5.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
gradio==5.16.2
gradio_client==1.7.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.0
idna==3.10
inflect==7.5.0
isodate==0.7.2
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kanjize==1.6.0
language-tags==1.2.0
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
microphone==0.3.1
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.27.1
networkx==3.4.2
numpy==2.2.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
phonemizer==3.3.0
pillow==11.1.0
pip==25.0.1
pluginmanager==0.4.1
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pylatexenc==2.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==7.1.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.22.3
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
segments==2.2.1
semantic-version==2.10.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
starlette==0.45.3
streamlit==1.42.1
SudachiDict-full==20250129
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
torch==2.6.0
torchaudio==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.49.0
typeguard==4.4.2
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.2
uvicorn==0.34.0
vexmessage==0.4.0
watchdog==6.0.0
websockets==14.2
yarl==1.18.3
zonos==0.1.0.dev0